revert split test binaries
